requirements.txt
langgraph
langchain-openai
python-dotenv
mcp
httpx
langchain-core
rich
pydantic
requests
pyyaml
supabase
markdown
openai
fastmcp
tavily-python
llmlingua
deepl
pymupdf4llm
google-api-python-client
google-auth-oauthlib
